requirements.txt
flask~=2.2.2
pymysql~=1.0.2
wtforms~=3.0.1
gunicorn
gevent
flask_wtf
pycrypto~=2.6.1